requirements.txt
boto3==1.34.162
langchain==0.2.16
langchain-aws==0.1.17
langchain-community==0.2.16
langchain-chroma==0.1.4
python-dotenv==1.0.1
streamlit==1.39.0
pandas==2.2.3
numpy==1.26.4
pypdf==4.3.1